Dashboard/requirements.txt
asgiref==3.8.1
Django==5.0.6
mysql-connector-python==9.0.0
mysqlclient==2.2.4
sqlparse==0.5.0
tzdata==2024.1
